Split report document projection tests
